🔥 로그인 관련 타입 파일과 유저 관련 타입 파일 통합
